implement CommentRepository with Spring Data Jpa
