+++ library/src/main/java/cyclone/otusspring/library/repository/datajpa/CommentJpaRepository.java
@@ -1,0 +1,17 @@
+package cyclone.otusspring.library.repository.datajpa;
+
+import cyclone.otusspring.library.model.Book;
+import cyclone.otusspring.library.model.Comment;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.List;
+
+public interface CommentJpaRepository extends JpaRepository<Comment, Long> {
+
+    List<Comment> findByBookOrderByDate(Book book);
+
+    List<Comment> findByBookBookIdOrderByDate(long bookId);
+
+    List<Comment> findByCommentatorOrderByDate(String commentator);
+
+}
