+++ library/src/test/java/cyclone/otusspring/library/repository/jpa/CommentRepositoryImplTest.java
@@ -13,8 +13,8 @@ import static cyclone.otusspring.library.TestData.*;
 import static org.assertj.core.api.Assertions.assertThat;
 
 @DataJpaTest
-@ComponentScan("cyclone.otusspring.library.repository.jpa")
-class CommentRepositoryJpaTest {
+@ComponentScan("cyclone.otusspring.library.repository")
+class CommentRepositoryImplTest {
 
     @Autowired
     CommentRepository commentRepository;
